mmc: host: sdhci-msm: skip power management entirely when SUSPEND_SKIP_SYNC is set
since it's completely broken when said config is set and even when all the functions which "waits till sync is done" (specially the workqueue ones) are converted to non-wait ones, it just keeps trying to suspend and resume everytime and fails to do so in the end, wasting power and precious suspend cycles, so just skip the whole suspend and resume functions during this state.

deep sleep drain remains unchanged, at min 20ma on average.